wrapper needs to be JSX
removing the < /> from the wrapper declaration in the last commit caused a new error:

Error: Element type is invalid: expected a string (for built-in components) or a class/function (for composite components) but got: undefined. You likely forgot to export your component from the file it's defined in. at invariant...

It appears that it needs to be in jsx format, not a direct reference to the react class in order to work. adding the < /> back in worked for me.